move GPSearchAlgorithm to hps
