+++ src/hps/search_algorithms/__init__.py
@@ -1,2 +1,2 @@
 from .search_algorithm import SearchAlgorithm, TrialPoint
-from .random_search_algorithm import RandomSearchAlgorithm
+from .classical import RandomSearchAlgorithm, GPSearchAlgorithm
