+++ src/hps/search_algorithms/classical/__init__.py
@@ -1,0 +1,3 @@
+from .random_search_algorithm import RandomSearchAlgorithm
+from .gp_search_algorithm import GPSearchAlgorithm
+
